Clear 12 build warnings + update CLAUDE.md default branch
Warnings (CS0618 obsolete API + CS4014 unawaited async) — all benign
on Revit 2025+ but worth clearing now that the build is otherwise green:

- SlopeAutoCorrector.cs:299 — ElementId.IntegerValue → .Value, switch
  cat to long, cast BuiltInCategory enum to long for comparison
- ConduitConsolidator.cs:210,230 — drop the (int) casts on the
  ElementId(...) calls; the new ElementId(Int64) constructor accepts
  the long ids from CableManifest.RouteTrayIds (List<long>) directly
- ConduitAutoRouteCommand.cs:182,213 — same Int64 ElementId fix
- SitePhotosTab.cs:549 — discard the DispatcherOperation with `_ = ...`
  to acknowledge the fire-and-forget BeginInvoke is intentional

CLAUDE.md — Branching section said `master` was the default branch, but
the repo's actual trunk is `main`. Updated. Added a note explicitly
warning against committing to local `main` (the source of the recurring
"14-commit divergence" we hit earlier this session).

+++ StingTools/Commands/Electrical/Routing/ConduitConsolidator.cs
@@ -207,7 +207,7 @@ namespace StingTools.Commands.Electrical.Routing
             {
                 if (c.RouteTrayIds == null || c.RouteTrayIds.Count == 0) continue;
                 rep = c;
-                firstConduitId = new ElementId((int)c.RouteTrayIds[0]);
+                firstConduitId = new ElementId(c.RouteTrayIds[0]);
                 break;
             }
             if (rep == null) return;
@@ -227,7 +227,7 @@ namespace StingTools.Commands.Electrical.Routing
                 {
                     try
                     {
-                        var el = doc.GetElement(new ElementId((int)lid));
+                        var el = doc.GetElement(new ElementId(lid));
                         if (el == null) continue;
                         doc.Delete(el.Id);
                         result.DeletedConduits++;
